+++ include/debug.h
@@ -1,7 +1,7 @@
 #ifndef DEBUG_H
 #define DEBUG_H
 
-#define IMU_DEBUG 1 // 2 for higher level 
+#define IMU_DEBUG 0 // 2 for higher level 
 #define GPS_DEBUG 0
 #define MQTT_DEBUG 1 
 #define CAN_DEBUG 1